chore: refactor fetch utils to http utils
