+++ utils/http/integrity.js
